+++ tests/testthat/helper-preproc.R
@@ -1,7 +1,7 @@
 prep_fun <- function(data, .by = NULL) {
   data |>
-    dplyr::group_by(dplyr::pick(dplyr::all_of(.by))) |>
-    dplyr::summarise(
+    group_by(pick(all_of(.by))) |>
+    summarise(
       nhit = mean(.data$nhit[.data$feedback == 1]),
       .groups = "drop"
     )